Check for modified data files once per minute
We still watch the local file system in the hope to catch changes
earlier. But we found out the hard way that file system notifications
seem to be unreliable on certain systems. We're not sure if the Linux
kernel fails to deliver file system notifications, or if they are
getting eaten up somewhere in Docker virtualization. Either way, we
now additionally poll for changes to be on the safe side.